:rocket: delete comment

+++ src/services/comments/deleteComments.service.ts
@@ -2,14 +2,20 @@ import AppDataSource from "../../data-source";
 import Comments from "../../entities/comments.entity";
 import AppError from "../../errors/appError";
 
-const deleteCommentsService = async (id: string) => {
+const deleteCommentsService = async (id: string, userId: string) => {
   const commentRepository = AppDataSource.getRepository(Comments);
-  const commentExists = commentRepository.findOneBy({
-    id
-  })
-  if (!commentExists){
+  const commentExists = await commentRepository.findOneBy({
+    id,
+  });
+
+  if (!commentExists) {
     throw new AppError("Comment does not exists", 404);
   }
+
+  if (commentExists!.user.id !== userId) {
+    throw new AppError("Unauthorized user", 401);
+  }
+
   await AppDataSource.createQueryBuilder()
     .delete()
     .from(Comments)
@@ -17,6 +23,6 @@ const deleteCommentsService = async (id: string) => {
     .execute();
 
   return;
-}
+};
 
 export default deleteCommentsService;
